refactor: change userid from integer to long in tests

+++ src/test/java/pl/edu/pk/student/kittysecurity/PasswordServiceTests.java
@@ -43,7 +43,7 @@ class PasswordServiceTests {
 
     private static final String JWT_TOKEN = "Bearer token";
     private static final String CLEANED_TOKEN = "token";
-    private static final Integer USER_ID = 42;
+    private static final Long USER_ID = 42L;
 
     private User sampleUser() {
         User user = new User();
@@ -74,7 +74,7 @@ class PasswordServiceTests {
                 .build();
 
         when(jwtService.extractUserId(CLEANED_TOKEN)).thenReturn(USER_ID.toString());
-        when(userRepo.findById(USER_ID)).thenReturn(Optional.of(sampleUser()));
+        when(userRepo.findById(Math.toIntExact(USER_ID))).thenReturn(Optional.of(sampleUser()));
         when(passwordEntryRepository.save(any(PasswordEntry.class))).thenAnswer(i -> i.getArguments()[0]);
         try (var mockedStatic = Mockito.mockStatic(JwtUtils.class)) {
             mockedStatic.when(() -> JwtUtils.cleanToken(JWT_TOKEN)).thenReturn(CLEANED_TOKEN);
@@ -93,7 +93,7 @@ class PasswordServiceTests {
     @Test
     void addPasswordByJwt_userNotFound_shouldThrow() {
         when(jwtService.extractUserId(CLEANED_TOKEN)).thenReturn(USER_ID.toString());
-        when(userRepo.findById(USER_ID)).thenReturn(Optional.empty());
+        when(userRepo.findById(Math.toIntExact(USER_ID))).thenReturn(Optional.empty());
         try (var mockedStatic = Mockito.mockStatic(JwtUtils.class)) {
             mockedStatic.when(() -> JwtUtils.cleanToken(JWT_TOKEN)).thenReturn(CLEANED_TOKEN);
 
@@ -107,7 +107,7 @@ class PasswordServiceTests {
         PasswordEntry entry = sampleEntry(user);
 
         when(jwtService.extractUserId(CLEANED_TOKEN)).thenReturn(USER_ID.toString());
-        when(userRepo.findById(USER_ID)).thenReturn(Optional.of(user));
+        when(userRepo.findById(Math.toIntExact(USER_ID))).thenReturn(Optional.of(user));
         when(passwordEntryRepository.findByUser(user)).thenReturn(List.of(entry));
         try (var mockedStatic = Mockito.mockStatic(JwtUtils.class)) {
             mockedStatic.when(() -> JwtUtils.cleanToken(JWT_TOKEN)).thenReturn(CLEANED_TOKEN);
@@ -131,7 +131,7 @@ class PasswordServiceTests {
         PasswordEntry entry = sampleEntry(user);
 
         when(jwtService.extractUserId(CLEANED_TOKEN)).thenReturn(USER_ID.toString());
-        when(userRepo.findById(USER_ID)).thenReturn(Optional.of(user));
+        when(userRepo.findById(Math.toIntExact(USER_ID))).thenReturn(Optional.of(user));
         when(passwordEntryRepository.findByUserAndEntryId(user, entry.getEntryId())).thenReturn(Optional.of(entry));
         try (var mockedStatic = Mockito.mockStatic(JwtUtils.class)) {
             mockedStatic.when(() -> JwtUtils.cleanToken(JWT_TOKEN)).thenReturn(CLEANED_TOKEN);
@@ -148,7 +148,7 @@ class PasswordServiceTests {
         User user = sampleUser();
 
         when(jwtService.extractUserId(CLEANED_TOKEN)).thenReturn(USER_ID.toString());
-        when(userRepo.findById(USER_ID)).thenReturn(Optional.of(user));
+        when(userRepo.findById(Math.toIntExact(USER_ID))).thenReturn(Optional.of(user));
         when(passwordEntryRepository.findByUserAndEntryId(user, 999L)).thenReturn(Optional.empty());
         try (var mockedStatic = Mockito.mockStatic(JwtUtils.class)) {
             mockedStatic.when(() -> JwtUtils.cleanToken(JWT_TOKEN)).thenReturn(CLEANED_TOKEN);
